Update README.md
Repository for the pre-bootcamp challenges and basic web development learning
This repository serves as a resource on our coding journey, encompassing a collection of pre-bootcamp challenges and foundational materials for learning basic web development. Designed to reinforce core programming concepts and delve into the fundamentals of web development. To explore, learn and help.